Update and rename ThorLabs Stack Splitter.ijm to ThorLabs Stack Splitter V2.ijm
fixed some major bugs require break the tiff readers. The Tiff reader in imageJ requires there to be a space somewhere in the directory when using the open command, e.g. C:/folder/directory/file.tiff will prevent the file from opening, but C:/folder/a directory/file.tiff will not. This new version adds a space to the file name to prevent this from happening and then removes it.